requirements/base.txt
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
confluent-kafka==1.3.0
dnspython==2.0.0
dotty-dict==1.2.1
elasticsearch==7.0.5
email-validator==1.1.2
fhir.resources==6.0.0b8
fhirpath @ git+https://github.com/arkhn/fhirpath@3d0227bb45073669e51cb2a56ba9733f710a0192
fhirspec==0.2.5
fhirstore @ git+https://github.com/arkhn/pyfhirstore@e11d997e911b7d2d07787dec5b431986cdc40bfc
idna==2.10
iniconfig==1.1.1
isodate==0.6.0
jsonschema==3.0.2
multidict==5.0.2
orjson==3.4.3
packaging==20.4
pluggy==0.13.1
py==1.9.0
pydantic==1.7.2
pymongo==3.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
python-dotenv==0.15.0
requests==2.24.0
setuptools-scm==4.1.2
six==1.15.0
toml==0.10.2
tqdm==4.36.1
urllib3==1.25.11
Werkzeug==1.0.1
yarl==1.6.3
zope.interface==5.2.0
